testing toast service from template
