+++ src/app/dashboard/dashboard.module.ts
@@ -1,9 +1,9 @@
-import { NgModule } from "@angular/core";
-import { CommonModule } from "@angular/common";
-import { FormsModule } from "@angular/forms";
-import { Routes, RouterModule } from "@angular/router";
-import { ChartsModule } from "ng2-charts";
-import { DashboardComponent } from "./dashboard.component";
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { Routes, RouterModule } from '@angular/router';
+import { ChartsModule } from 'ng2-charts';
+import { DashboardComponent } from './dashboard.component';
 import { SalesGraphComponent } from './dashboard-components/sales-graph/sales-graph.component';
 import { VisitGraphComponent } from './dashboard-components/visit-graph/visit-graph.component';
 import { WebsiteGraphComponent } from './dashboard-components/website-graph/website-graph.component';
@@ -12,23 +12,31 @@ import { ContactsComponent } from './dashboard-components/contacts/contacts.comp
 import { ProjectOfMonthComponent } from './dashboard-components/project-of-month/project-of-month.component';
 
 const routes: Routes = [
-  {
-    path: "",
-    data: {
-      title: "Dashboard",
-      urls: [{ title: "Dashboard", url: "/dashboard" }, { title: "Dashboard" }],
-    },
-    component: DashboardComponent,
-  },
+	{
+		path: '',
+		data: {
+			title: 'Dashboard',
+			urls: [{ title: 'Dashboard', url: '/dashboard' }, { title: 'Dashboard' }],
+		},
+		component: DashboardComponent,
+	},
 ];
 
 @NgModule({
-  imports: [
-    FormsModule,
-    CommonModule,
-    RouterModule.forChild(routes),
-    ChartsModule,
-  ],
-  declarations: [DashboardComponent, SalesGraphComponent, VisitGraphComponent, WebsiteGraphComponent, CardsComponent, ContactsComponent, ProjectOfMonthComponent],
+	imports: [
+		FormsModule,
+		CommonModule,
+		RouterModule.forChild(routes),
+		ChartsModule,
+	],
+	declarations: [
+		DashboardComponent,
+		SalesGraphComponent,
+		VisitGraphComponent,
+		WebsiteGraphComponent,
+		CardsComponent,
+		ContactsComponent,
+		ProjectOfMonthComponent,
+	],
 })
 export class DashboardModule {}
